+++ src/apps/popup/pages/download-secret-keys/utils.ts
@@ -1,5 +1,3 @@
-import { CLPublicKey, CLPublicKeyTag, decodeBase64, Keys } from 'casper-js-sdk';
-
 export function downloadFile(content: Blob, filename: string): void {
   const url = window.URL.createObjectURL(content);
   const downloadFileLink = document.createElement('a');
@@ -8,31 +6,4 @@ export function downloadFile(content: Blob, filename: string): void {
   downloadFileLink.setAttribute('download', filename);
   downloadFileLink.click();
   downloadFileLink.remove();
-}
-
-export function makeSecretKeyFileContent(
-  publicKey: string,
-  secretKey: string
-): string {
-  const clPublicKey = CLPublicKey.fromHex(publicKey);
-  const decodedSecretKey = decodeBase64(secretKey);
-
-  let keyPair;
-  switch (clPublicKey.tag) {
-    case CLPublicKeyTag.ED25519:
-      keyPair = new Keys.Ed25519({
-        publicKey: clPublicKey.value(),
-        secretKey: decodedSecretKey
-      });
-      break;
-
-    case CLPublicKeyTag.SECP256K1:
-      keyPair = new Keys.Secp256K1(clPublicKey.value(), decodedSecretKey);
-      break;
-
-    default:
-      throw Error('Unknown Signature type.');
-  }
-
-  return keyPair.exportPrivateKeyInPem();
 }
